DDCNL-1725 | PostcodeLookupController created
