+++ app/views/personaldetails/cannotUseService.scala.html
@@ -34,7 +34,7 @@ pageName = Some(messages("label.your_personal_tax_account"))
             </header>
 
     <p class="lede margin-bottom">@messages("label.this_is_because_you_do_not_live_in_the_uk")</p>
-    <p class="margin-top">@messages("label.if_you_do_live_in_the_uk_text_before_link") <a class="ga-track-anchor-click" data-ga-event-category="link - click" data-ga-event-action="You cannot use this service to update your address" data-ga-event-label="change your address" href="@controllers.routes.AddressController.showPostcodeLookupForm(typ)">@messages("label.if_you_do_live_in_the_uk_link_text")</a>.</p>
+    <p class="margin-top">@messages("label.if_you_do_live_in_the_uk_text_before_link") <a class="ga-track-anchor-click" data-ga-event-category="link - click" data-ga-event-action="You cannot use this service to update your address" data-ga-event-label="change your address" href="@address.routes.PostcodeLookupController.onPageLoad(typ)">@messages("label.if_you_do_live_in_the_uk_link_text")</a>.</p>
 
     <p>@messages("label.if_you_do_not_live_in_the_uk_text_before_link") <a class="ga-track-anchor-click" data-ga-event-category="link - click" data-ga-event-action="You cannot use this service to update your address" data-ga-event-label="fill in the ‘Notification of a change in personal details’ form to update your address (opens in a new window or tab)" href="@configDecorator.updateAddressShortFormUrl" target="_blank" rel="noopener noreferrer">@messages("label.if_you_do_not_live_in_the_uk_link_text") @messages("label.opens_in_a_new_window")</a>.</p>
   </div>
