+++ conf/prod.routes
@@ -43,8 +43,8 @@ POST        /personal-account/your-address/residency-choice                     
 GET         /personal-account/your-address/:typ/do-you-live-in-the-uk                  controllers.address.InternationalAddressChoiceController.onPageLoad(typ: AddrType)
 POST        /personal-account/your-address/:typ/do-you-live-in-the-uk                  controllers.address.InternationalAddressChoiceController.onSubmit(typ: AddrType)
 GET         /personal-account/your-address/:typ/cannot-use-the-service                 controllers.AddressController.cannotUseThisService(typ: AddrType)
-GET         /personal-account/your-address/:typ/find-address                           controllers.AddressController.showPostcodeLookupForm(typ: AddrType)
-POST        /personal-account/your-address/:typ/find-address                           controllers.AddressController.processPostcodeLookupForm(typ: AddrType)
+GET         /personal-account/your-address/:typ/find-address                           controllers.address.PostcodeLookupController.onPageLoad(typ: AddrType)
+POST        /personal-account/your-address/:typ/find-address                           controllers.address.PostcodeLookupController.onSubmit(typ: AddrType)
 GET         /personal-account/your-address/:typ/select-address                         controllers.AddressController.showAddressSelectorForm(typ: AddrType)
 POST        /personal-account/your-address/:typ/select-address                         controllers.AddressController.processAddressSelectorForm(typ: AddrType)
 GET         /personal-account/your-address/:typ/edit-address                           controllers.AddressController.showUpdateAddressForm(typ: AddrType)
