+++ test/services/AddressLookupServiceSpec.scala
@@ -19,7 +19,7 @@ package services
 import com.codahale.metrics.Timer
 import com.kenshoo.play.metrics.Metrics
 import config.ConfigDecorator
-import models.addresslookup.RecordSet
+import models.addresslookup.{AddressLookupErrorResponse, AddressLookupSuccessResponse, AddressLookupUnexpectedResponse, RecordSet}
 import org.mockito.Matchers._
 import org.mockito.Mockito._
 import org.scalatestplus.mockito.MockitoSugar
